Fix flow identity resolution and address code review findings
Flow Explorer returned "matches N source files" even when the caller sent the
exact source-relative path: the match was a flat OR over path/basename/stem, so
a sibling sharing a basename matched via stem and forced the ambiguity branch.
Since Dependency Health flags exactly those programs as ambiguous and the UI
answers by sending the full path, the two surfaces contradicted each other.
Resolution is now tiered — path, then basename, then stem — and the ambiguity
list is rendered as buttons that re-query the chosen path instead of inert text.

Also from review:
- The job picker was built from the filtered result, so choosing a job erased
  every other option. ServiceChainSnapshot now carries the unfiltered domain.
- The scratch-path rule existed five times with three different bodies; two
  missed .convert- and two used a "/.convert-" substring that never matches on
  Windows. Consolidated into SourceTypeRegistry.IsScratchPath.
- BuildServiceChainMermaid returned a string while writing two more fields into
  its argument; it now returns all three and a named ApplyMermaid assigns them.
- CancellationToken was accepted and ignored by all three graph endpoints.
- SourcePathHelper and ProgramFactsArtifactLocator are public, so the portal no
  longer needs InternalsVisibleTo over the whole CLI assembly.
- Moved DTOs into McpChatWeb/Models to match the one-or-two-types-per-file layout.

Verified against a live backend over a temporary estate with duplicate
basenames, two JCL jobs and a .convert- directory: exact paths resolve, the
job and program filters round-trip, and scratch copies stay excluded.

Copilot-Session: b888b510-b445-4ff7-aa16-29391494baa6
